Add MQTT flag to control MQTT functionality
Introduced a new `mqtt` boolean member to `DetectorEnvironment`,
`camera_settings`, and `device_settings` classes to control
whether MQTT functionality is enabled. Updated `init_detectors_environment`
to initialize the `mqtt` member from `camera_setting`. Modified
`main` to check the `mqtt` flag before connecting to the MQTT broker.
Added parsing of the `mqtt` field in `camera_settings::parse` and
`device_settings::parse` functions. Default value for `mqtt` is
currently `true`, with a note to change it to `false` in the future.